Rename Components so that they're capitalized
