+++ public/style.css
@@ -67,7 +67,7 @@ select{
 }
 
 
-/* User Dashboard  --- user-home.js
+/* User Dashboard  --- UserHome.js
 */
 /* navbar */
 .logout-btn{
